Closes #3
Also upgrading a couple dependencies
Fixing election view when not voting to not show pointer cursor.
Specifying that the only disabled submit button should be the voting one
(so that we can submit the voter files).
Storing voters for an election. Now tokens should be generated only upon
election opening.
We also now integrate with the identity system for loging in. Log in is
currently only useful for admins (me at this point) to decide who can
register voters and who cannot. This should change in the future so that
logged in users can create new elections for which they become admins.