Closes #8
Improving the migrations. Adding a singleton clock process to trigger
emails to voters upon reaching election start times.
This is not ideal. It shouldn't run the actual work but rather queue up
work for some other worker to pick but I don't want to add the
complexity of a queueing system and workers.

Also changed the log4j properties to only track info logging
information in preprod and prod while being able to override whenever
desired.