Fixing issue in session management
To improve on #7 and to better handle session management added in #3,
sessions have to be fully set when they're messed with.
This commit makes it so that if you once specify your locale using
request params, it gets persisted in your session. If you remove
your session (via logout), we'll ensure the redirect request maintains
the locale information so you stay in the same language you were before.
If you log in, we'll merge the log in information into your existing
session along with your locale.